it works. the problem was that the wasm was not being regenerated
because of some typos in the c code

The character encoding of the HTML document was not declared. The document will render with garbled text in some browser configurations if the document contains characters from outside the US-ASCII range. The character encoding of the page must be declared in the document or in the transfer protocol. index2.html
The Components object is deprecated. It will soon be removed. index2.html
15